Refactor split config methods and schema
