+++ src/client/bitrise_client.ts
@@ -2,7 +2,7 @@ import { AxiosInstance } from 'axios'
 import { minBy } from 'lodash'
 import { Artifact, createAxios, CustomReportArtifact } from './client'
 import { minimatch } from 'minimatch'
-import { CustomReportConfig } from '../config/config'
+import { CustomReportConfig } from '../config/schema'
 import { ArgumentOptions } from '../arg_options'
 import { Logger } from 'tslog'
