+++ src/runner/bitrise_runner.ts
@@ -5,7 +5,7 @@ import { BitriseAnalyzer } from "../analyzer/bitrise_analyzer"
 import { ArgumentOptions } from "../arg_options"
 import { BitriseClient } from "../client/bitrise_client"
 import { BitriseConfig, parseConfig } from "../config/bitrise_config"
-import { ValidatedYamlConfig } from "../config/validator"
+import { ValidatedYamlConfig } from "../config/config"
 import { createCustomReportCollection, CustomReportCollection } from "../custom_report_collection"
 import { CompositExporter } from "../exporter/exporter"
 import { LastRunStore } from "../last_run_store"
